Rename jax_runner_v2 to jax_runner.
